pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pants-sls-distribution"
dynamic = ["version"]
description = "Pants plugin for SLS-compatible Python service distributions with Apollo manifest generation"
readme = "README.md"
requires-python = ">=3.11,<4"
license = "Apache-2.0"
authors = [{ name = "jaymd96" }]
keywords = ["pants", "sls", "palantir", "apollo", "packaging"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20.0",
    "pants-docker-generator>=0.1.0",
    "pants-apollo-hub>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pantsbuild.pants>=2.24.0",
    "pantsbuild.pants.testutil>=2.24.0",
]

[tool.hatch.version]
path = "src/pants_sls_distribution/__about__.py"
pattern = '__version__ = "(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
packages = ["src/pants_sls_distribution"]

[tool.hatch.build.targets.sdist]
include = ["src/pants_sls_distribution", "src/pants_sls_distribution/schemas"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]

[tool.hatch.envs.default]
features = ["dev"]

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
test-cov = "pytest --cov=pants_sls_distribution {args}"
